* imaging: Fix adding a client to a profile
Since the specification tells that a client cannot be
in multiple profiles we check the ProfilesResults table
in the dyngroup database to check that the client is not
in any already created profile.

The code was checking the imaging database and returning
any client that has a bootmenu. The problem is that any
client has a bootmenu. So no client could be added to any
profile.

A typo in the method name was fixed also.